+++ tests/test_logging.py
@@ -44,6 +44,8 @@ from click_extra.pytest import (
     default_debug_uncolored_logging,
 )
 
+from .conftest import skip_windows_colors
+
 
 def test_level_default_order():
     assert tuple(LOG_LEVELS) == ("CRITICAL", "ERROR", "WARNING", "INFO", "DEBUG")
@@ -129,6 +131,7 @@ def test_unrecognized_verbosity(invoke, cmd_decorator, cmd_type):
     )
 
 
+@skip_windows_colors
 @pytest.mark.parametrize(
     "cmd_decorator",
     # Skip click extra's commands, as verbosity option is already part of the default.